aws_iotcore_pubsub/requirements.txt
awscrt==0.16.0
awsiotsdk==1.12.0
